--- app/src/main/res/xml/default_browser_preferences.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- This Source Code Form is subject to the terms of the Mozilla Public
-   - License, v. 2.0. If a copy of the MPL was not distributed with this
-   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <org.mozilla.fenix.settings.DefaultBrowserPreference
-        android:key="@string/pref_key_make_default_browser"
-        android:title="@string/preferences_set_as_default_browser"
-        app:iconSpaceReserved="false" />
-</PreferenceScreen>
